+++ src/locale.go
@@ -10,7 +10,7 @@ import (
 	"time"
 )
 
-const appVersion = "1.0.0"
+const appVersion = "1.1.0"
 const appID = "Astral"
 const appNameZH = "七曜工作法"
 const appNameEN = "Astral Rhythm"